Refactor SendScreen
