+++ apps/daimo-mobile/src/view/screen/receive/RequestScreen.tsx
@@ -2,11 +2,11 @@ import { tokenMetadata } from "@daimo/contract";
 import { useCallback, useState } from "react";
 import { Alert, ScrollView, Share, StyleSheet } from "react-native";
 
-import { assert } from "../../logic/assert";
-import { useAccount } from "../../model/account";
-import { ButtonBig } from "../shared/Button";
-import { AmountInput } from "../shared/Input";
-import { color } from "../shared/style";
+import { assert } from "../../../logic/assert";
+import { useAccount } from "../../../model/account";
+import { ButtonBig } from "../../shared/Button";
+import { AmountInput } from "../../shared/Input";
+import { color } from "../../shared/style";
 
 export default function RequestScreen() {
   const [account] = useAccount();
